New sansred.steps.param_ratio module
This module takes in two parameter object and takes their ratio abiding by the chosen alignment. This is intended to be used with calculating transmission ratios used in background subtraction. This module is messier than is probably necessary, but attempts to handle the edge cases where zero or only one of the parameters is specified.