Added sansred.steps.moments_fit
The moments function was failing to correctly identify the beam center when the detector was offset. This new method fits a 2D gaussian to the image which is much more computationally intensive but actually produces the correct result more reliably. I've added a dependency to lmfit for this method which could easily be replaced by any fitting library (homegrown or otherwise).